pyproject.toml
[project]
name = "webapp-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=11.2.1",
    "streamlit>=1.45.1",
    "streamlit-image-select>=0.6.0",
    "torch>=2.7.0",
    "transformers>=4.52.3",
]
